healthcare insurance/requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Document Processing
datasets>=2.12.0
Pillow>=9.5.0
opencv-python>=4.7.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# Layout & Vision Models
layoutlm>=0.1.0
timm>=0.9.2

# API & Web Service
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
kaggle>=1.5.0

# Validation & Regex
phonenumbers>=8.13.0
validators>=0.20.0

# Azure/Power BI Integration (optional)
pyodbc>=4.0.39
azure-storage-blob>=12.16.0 